:sparkles: update Kotlin logo path
Moves the kotlin logo image file to a new folder called 'attach' instead of
'assets'. This is where Simiki gets attachments from. While we could override
this to read from the 'assets' folder, the Simiki documentation discourages us
from doing so. Not sure why, but I won't argue with Tankywoo-san.