:wrench: stop tracking simiki output folder
The output folder is updated everytime we generate new content. If running
Simiki in watch mode, this folder is updated every time a markdown file
changes. It can become noisy to commit all these changes. Additionally, it
seems that other Simiki users don't track output folder, so it's probably not
meant to be tracked in the first place.